refactor: moved database files into separate directory and updated readme
